@@ -1,6 +1,6 @@
 buildscript {
     ext {
-        kotlin_version = '1.4.32'
+        kotlin_version = '1.5.0'
         about_libraries = '8.8.5'
         hilt_version = "2.35.1"
     }
