@@ -52,9 +52,6 @@ class MoreFragment : BaseFragment<FragmentMoreBinding>(R.layout.fragment_more), 
     override val luckyNumberRes: Pair<String, Drawable?>?
         get() = context?.run { getString(R.string.lucky_number_title) to getCompatDrawable(R.drawable.ic_more_lucky_number) }
 
-    override val mobileDevicesRes: Pair<String, Drawable?>?
-        get() = context?.run { getString(R.string.mobile_devices_title) to getCompatDrawable(R.drawable.ic_more_mobile_devices) }
-
     override val conferencesRes: Pair<String, Drawable?>?
         get() = context?.run { getString(R.string.conferences_title) to getCompatDrawable(R.drawable.ic_more_conferences) }
 
@@ -63,6 +60,9 @@ class MoreFragment : BaseFragment<FragmentMoreBinding>(R.layout.fragment_more), 
 
     override val schoolAndTeachersRes: Pair<String, Drawable?>?
         get() = context?.run { getString(R.string.schoolandteachers_title) to getCompatDrawable((R.drawable.ic_more_schoolandteachers)) }
+
+    override val mobileDevicesRes: Pair<String, Drawable?>?
+        get() = context?.run { getString(R.string.mobile_devices_title) to getCompatDrawable(R.drawable.ic_more_mobile_devices) }
 
     override val settingsRes: Pair<String, Drawable?>?
         get() = context?.run { getString(R.string.settings_title) to getCompatDrawable(R.drawable.ic_more_settings) }
@@ -112,10 +112,6 @@ class MoreFragment : BaseFragment<FragmentMoreBinding>(R.layout.fragment_more), 
         (activity as? MainActivity)?.pushView(LuckyNumberFragment.newInstance())
     }
 
-    override fun openMobileDevicesView() {
-        (activity as? MainActivity)?.pushView(MobileDeviceFragment.newInstance())
-    }
-
     override fun openSchoolAnnouncementView() {
         (activity as? MainActivity)?.pushView(SchoolAnnouncementFragment.newInstance())
     }
@@ -126,6 +122,10 @@ class MoreFragment : BaseFragment<FragmentMoreBinding>(R.layout.fragment_more), 
 
     override fun openSchoolAndTeachersView() {
         (activity as? MainActivity)?.pushView(SchoolAndTeachersFragment.newInstance())
+    }
+
+    override fun openMobileDevicesView() {
+        (activity as? MainActivity)?.pushView(MobileDeviceFragment.newInstance())
     }
 
     override fun openSettingsView() {
